prepare for 64bit host-system build
Since the fact that the root of this project (modula2) has never seen
64bits and therefore expects that things like 'unsigned long' always
have 32bit. But at todays 64bit system things did change.

For this we convert all datatypes to be very precise in their size.

unsigned long	--> 	uint32_t
long		-->	int32_t
unsigned char	-->	uint8_t
char		-->	int8_t
unsigned short	-->	uint16_t
short		-->	int16_t

+++ src/Execlogin.h
@@ -22,7 +22,7 @@ struct Execlogin_tEXEC {
    X2C_pCHAR * args;
 };
 
-extern long Execlogin_StartLogin(struct Execlogin_tEXEC *);
+extern int32_t Execlogin_StartLogin(struct Execlogin_tEXEC *);
 
 
 extern void Execlogin_BEGIN(void);
